backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis[hiredis]==5.0.4
minio==7.2.7
langchain==0.2.1
langchain-community==0.2.1
langchain-ollama==0.1.1
pypdf==4.2.0
slowapi==0.1.9
anthropic==0.28.0
